refactor: extract header related components and event handlers
